requirements.txt
# Core dependencies
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.4
croniter>=2.0.0
pytz>=2024.1

# Logging
structlog>=24.1.0

# Configuration
pyyaml>=6.0.1

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Retries
tenacity>=8.2.3

# Monitoring
prometheus-client>=0.19.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# CLI
click>=8.1.7
rich>=13.7.0

# Database migrations (optional)
alembic>=1.13.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.1

# Code quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1
pylint>=3.0.3

# Type stubs
types-requests>=2.31.0
types-pyyaml>=6.0.12

# Development
ipython>=8.18.1
ipdb>=0.13.13
